refactor structure tests
