+++ klass-api/src/test/java/no/ssb/klass/api/migration/dataintegrity/corresponds/KlassApiClassificationCorrespondsJsonTest.java
@@ -1,36 +1,14 @@
-package no.ssb.klass.api.migration.dataintegrity;
+package no.ssb.klass.api.migration.dataintegrity.corresponds;
 
-import io.restassured.response.Response;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 import java.time.Instant;
-import java.util.HashMap;
-import java.util.Map;
 
 import static no.ssb.klass.api.migration.MigrationTestConstants.*;
 import static no.ssb.klass.api.migration.MigrationTestUtils.*;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-public class KlassApiClassificationCorrespondsTest extends AbstractKlassApiDataIntegrityTest {
-
-
-    static Map<String, Object> paramsTargetIdAndDateFrom = new HashMap<>();
-
-    static Integer targetClassificationIdValue;
-    static String fromDate;
-
-    Response sourceResponse;
-    Response targetResponse;
-
-    @BeforeAll
-    static void beforeAllCorrespondence() {
-        fromDate = "2020-01-01";
-        targetClassificationIdValue = 103;
-        paramsTargetIdAndDateFrom.put(TARGET_CLASSIFICATION_ID, targetClassificationIdValue);
-        paramsTargetIdAndDateFrom.put(FROM, fromDate);
-
-    }
+public class KlassApiClassificationCorrespondsJsonTest extends AbstractKlassApiCorrespondsTest {
 
     @Test
     void getOneClassificationCorrespondence(){
@@ -51,9 +29,5 @@ public class KlassApiClassificationCorrespondsTest extends AbstractKlassApiDataI
         else{
             validateList(sourceResponse, targetResponse, CORRESPONDENCE_ITEMS);
         }
-    }
-
-    String getCorrespondsPath(Integer id) {
-        return CLASSIFICATIONS_PATH + "/"+ id + "/" + CORRESPONDS;
     }
 }
