+++ klass-api/src/test/java/no/ssb/klass/api/migration/dataintegrity/families/AbstractKlassApiFamiliesTest.java
@@ -1,0 +1,19 @@
+package no.ssb.klass.api.migration.dataintegrity.families;
+
+import no.ssb.klass.api.migration.dataintegrity.AbstractKlassApiDataIntegrityTest;
+
+import java.util.stream.IntStream;
+import java.util.stream.Stream;
+
+import static no.ssb.klass.api.migration.MigrationTestConstants.CLASSIFICATION_FAMILIES;
+
+public class AbstractKlassApiFamiliesTest extends AbstractKlassApiDataIntegrityTest {
+
+    static Stream<Integer> rangeProviderClassificationFamilyIds() {
+        return IntStream.rangeClosed(0, 30).boxed();
+    }
+
+    String getClassificationFamilyByIdPath(Integer id) {
+        return "/" + CLASSIFICATION_FAMILIES + "/" + id;
+    }
+}
