+++ klass-api/src/test/java/no/ssb/klass/api/migration/dataintegrity/families/KlassApiClassificationFamiliesTest.java
@@ -1,7 +1,6 @@
-package no.ssb.klass.api.migration.dataintegrity;
+package no.ssb.klass.api.migration.dataintegrity.families;
 
 import io.restassured.response.Response;
-import no.ssb.klass.api.migration.dataintegrity.codes.KlassApiClassificationCodesJsonTest;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
@@ -12,7 +11,7 @@ import static no.ssb.klass.api.migration.MigrationTestConstants.*;
 import static no.ssb.klass.api.migration.MigrationTestUtils.*;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-public class KlassApiClassificationFamiliesTest extends KlassApiClassificationCodesJsonTest {
+public class KlassApiClassificationFamiliesTest extends AbstractKlassApiFamiliesTest {
 
     static Map<String, Object> paramsSsbSection = new HashMap<>();
