+++ klass-api/src/test/java/no/ssb/klass/api/migration/dataintegrity/variant/KlassApiClassificationVariantAtTest.java
@@ -1,36 +1,14 @@
-package no.ssb.klass.api.migration.dataintegrity;
+package no.ssb.klass.api.migration.dataintegrity.variant;
 
-import io.restassured.response.Response;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 import java.time.Instant;
-import java.util.HashMap;
-import java.util.Map;
 
-import static no.ssb.klass.api.migration.MigrationTestConstants.*;
 import static no.ssb.klass.api.migration.MigrationTestConstants.CODES;
 import static no.ssb.klass.api.migration.MigrationTestUtils.*;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-public class KlassApiClassificationVariantAtTest extends AbstractKlassApiDataIntegrityTest{
-
-    static String variantNameId84 = "Klimagasser";
-    static String variantDateId84 = "2015-01-01";
-
-    static Map<String, Object> paramsVariantDate = new HashMap<>();
-
-    Response sourceResponse;
-    Response targetResponse;
-
-
-    @BeforeAll
-    static void beforeAllVariant() {
-        paramsVariantDate.put(VARIANT_NAME, variantNameId84);
-        paramsVariantDate.put(DATE, variantDateId84);
-
-    }
-
+public class KlassApiClassificationVariantAtTest extends AbstractKlassApiVariantTest {
 
     @Test
     void getOneClassificationVariantAt(){
@@ -51,9 +29,5 @@ public class KlassApiClassificationVariantAtTest extends AbstractKlassApiDataInt
         else{
             validateList(sourceResponse, targetResponse, CODES);
         }
-    }
-
-    String getVariantAtPath(Integer id) {
-        return CLASSIFICATIONS_PATH + "/"+ id + "/" + VARIANT_AT;
     }
 }
